@@ -1,6 +1,6 @@
 {
-  "name": "@quocanhbk/express-cassandra",
-  "version": "6.0.0",
+  "name": "@quocanhbk17/express-cassandra",
+  "version": "0.1.8",
   "description": "Nest - modern, fast, powerful node.js web framework (@express-cassandra)",
   "author": {
     "name": "Fahim Rahman",
@@ -10,16 +10,17 @@
   "license": "MIT",
   "scripts": {
     "build": "rm -rf dist && tsc -p tsconfig.json",
+    "watch": "tsc -w",
     "precommit": "lint-staged",
     "prepublish:npm": "npm run build",
     "publish:npm": "npm publish --access public",
     "prepublish:next": "npm run build",
     "publish:next": "npm publish --access public --tag next",
-    "lint": "tslint -p tsconfig.json -c tslint.json"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
-    "url": "git@github.com:ifaim/nestjs-express-cassandra.git"
+    "url": "git+ssh://git@github.com/ifaim/nestjs-express-cassandra.git"
   },
   "homepage": "https://github.com/ifaim/nestjs-express-cassandra",
   "peerDependencies": {
@@ -32,15 +33,19 @@
   "devDependencies": {
     "@nestjs/common": "^10.4.6",
     "@nestjs/core": "^10.4.6",
+    "@types/jest": "^29.5.14",
+    "@types/merge-deep": "^3.0.3",
     "@types/node": "10.12.18",
     "express-cassandra": "^2.9.1",
     "husky": "3.0.9",
+    "jest": "^29.7.0",
     "lint-staged": "9.4.2",
     "prettier": "1.18.2",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
+    "ts-jest": "^29.2.5",
     "tslint": "5.20.0",
-    "typescript": "3.6.4"
+    "typescript": "^5.6.3"
   },
   "lint-staged": {
     "*.ts": [
@@ -61,5 +66,22 @@
     "node",
     "nodejs",
     "javascript"
-  ]
+  ],
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "ts"
+    ],
+    "rootDir": ".",
+    "testRegex": "(/tests/.*|(\\.|/)(test|spec))\\.ts$",
+    "transform": {
+      "^.+\\.(t|j)s$": "ts-jest"
+    },
+    "collectCoverageFrom": [
+      "**/*.(t|j)s"
+    ],
+    "coverageDirectory": "../coverage",
+    "testEnvironment": "node"
+  }
 }
