@@ -10,26 +10,27 @@
   },
   "scripts": {
     "format": "prettier --write \"src/**/*.ts\"",
-    "start": "ts-node -r tsconfig-paths/register src/main.ts",
-    "start:dev": "nodemon",
-    "start:debug": "nodemon --config nodemon-debug.json",
+    "start": "nest start",
+    "start:dev": "nest start --watch",
+    "start:debug": "nest start --debug --watch",
     "prestart:prod": "rimraf dist && tsc",
     "start:prod": "node dist/main.js",
     "lint": "tslint -p tsconfig.json -c tslint.json",
     "test": "jest"
   },
   "dependencies": {
-    "@iaminfinity/express-cassandra": "^5.0.2",
-    "@nestjs/common": "^5.5.0",
-    "@nestjs/core": "^5.5.0",
-    "@nestjs/microservices": "^5.5.0",
-    "@nestjs/testing": "^5.5.0",
-    "@nestjs/websockets": "^5.5.0",
-    "class-transformer": "^0.2.0",
-    "class-validator": "^0.9.1",
-    "reflect-metadata": "^0.1.12",
-    "rxjs": "^6.0.0",
-    "typescript": "^3.0.1"
+    "@nestjs/common": "^10.4.6",
+    "@nestjs/core": "^10.4.6",
+    "@nestjs/microservices": "^10.4.6",
+    "@nestjs/platform-express": "^10.4.6",
+    "@nestjs/testing": "^10.4.6",
+    "@nestjs/websockets": "^10.4.6",
+    "@quocanhbk17/express-cassandra": "^0.1.6",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.1",
+    "reflect-metadata": "^0.2.2",
+    "rxjs": "^7.8.1",
+    "typescript": "^5.6.3"
   },
   "devDependencies": {
     "@types/express": "^4.16.0",
