@@ -1,5 +1,7 @@
 export * from './column.decorator';
-export * from './entity.decorator';
-export * from './entity-repository.decorator';
-export * from './listeners';
 export * from './entity-method.decorator';
+export * from './entity-repository.decorator';
+export * from './entity.decorator';
+export * from './listeners';
+export * from './user-defined-type-column.decorator';
+export * from './user-defined-type.decorator';
